CKS-344 Refactor link re-writing
