+++ gatsby/plugins/gatsby-source-cks/src/link-rewriter/types.ts
